+++ Runtime/CameraSubsystem/XRCameraSubsystem.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using Unity.Collections;
 
 namespace UnityEngine.XR.ARSubsystems
@@ -209,6 +210,18 @@ namespace UnityEngine.XR.ARSubsystems
                 Allocator allocator)
             {
                 return new NativeArray<XRTextureDescriptor>(0, allocator);
+            }
+
+            /// <summary>
+            /// Method to be implemented by the provider to get the enabled and disabled shader keywords for the
+            /// material.
+            /// </summary>
+            /// <param name="enabledKeywords">The keywords to enable for the material.</param>
+            /// <param name="disabledKeywords">The keywords to disable for the material.</param>
+            public virtual void GetMaterialKeywords(out List<string> enabledKeywords, out List<string> disabledKeywords)
+            {
+                enabledKeywords = null;
+                disabledKeywords = null;
             }
 
             /// <summary>
@@ -607,6 +620,14 @@ namespace UnityEngine.XR.ARSubsystems
         /// <c>true</c> if camera permission has been granted. Otherwise, <c>false</c>.
         /// </value>
         public bool permissionGranted => m_Provider.permissionGranted;
+
+        /// <summary>
+        /// Get the enabled and disabled shader keywords for the material.
+        /// </summary>
+        /// <param name="enabledKeywords">The keywords to enable for the material.</param>
+        /// <param name="disabledKeywords">The keywords to disable for the material.</param>
+        public void GetMaterialKeywords(out List<string> enabledKeywords, out List<string> disabledKeywords)
+            => m_Provider.GetMaterialKeywords(out enabledKeywords, out disabledKeywords);
 
         /// <summary>
         /// Attempt to get the latest camera image. This provides directly access to the raw pixel data, as well as