+++ Runtime/RaycastSubsystem/XRRaycastSubsystem.cs
@@ -10,7 +10,7 @@ namespace UnityEngine.XR.ARSubsystems
     /// This abstract class should be implemented by an XR provider and instantiated using the <c>SubsystemManager</c>
     /// to enumerate the available <see cref="XRRaycastSubsystemDescriptor"/>s.
     /// </remarks>
-    public abstract class XRRaycastSubsystem : XRSubsystem<XRRaycastSubsystemDescriptor>
+    public abstract class XRRaycastSubsystem : TrackingSubsystem<XRRaycast, XRRaycastSubsystemDescriptor>
     {
         /// <summary>
         /// Constructor. Do not invoke directly; use the <c>SubsystemManager</c>
@@ -33,6 +33,53 @@ namespace UnityEngine.XR.ARSubsystems
         /// Destroys the subsystem.
         /// </summary>
         protected sealed override void OnDestroyed() => m_Provider.Destroy();
+
+        /// <summary>
+        /// Get the changes (arrays of added, updated and removed) to the tracked raycasts since the last call to this
+        /// method.
+        /// </summary>
+        /// <param name="allocator">An [allocator](https://docs.unity3d.com/ScriptReference/Unity.Collections.Allocator.html)
+        /// to use for the returned container.</param>
+        /// <returns>The set of changes since the last call to this method. The caller owns the data and is responsible
+        /// for calling <see cref="TrackableChanges{T}.Dispose"/> on it.</returns>
+        public override TrackableChanges<XRRaycast> GetChanges(Allocator allocator)
+        {
+            var changes = m_Provider.GetChanges(XRRaycast.defaultValue, allocator);
+#if DEVELOPMENT_BUILD || UNITY_EDITOR
+            m_ValidationUtility.ValidateAndDisposeIfThrown(changes);
+#endif
+            return changes;
+        }
+
+        /// <summary>
+        /// Attempts to add a new persistent raycast. The raycast will be updated automatically until
+        /// this subsystem is stopped or destroyed, or the the raycast is removed with
+        /// <see cref="RemoveRaycast(TrackableId)"/>.
+        /// </summary>
+        /// <param name="screenPoint">A point on the screen, in normalized screen coorindates (0, 0)..(1, 1)</param>
+        /// <param name="estimatedDistance">The estimated distance to the raycast target. For example, an average
+        /// human height might be used to estimate the distance to the floor.</param>
+        /// <param name="raycast">The newly added raycast. All spatial data is relative to the session origin.</param>
+        /// <returns>`True` if the raycast was successfully added, or `false` otherwise.</returns>
+        public bool TryAddRaycast(Vector2 screenPoint, float estimatedDistance, out XRRaycast raycast) => m_Provider.TryAddRaycast(screenPoint, estimatedDistance, out raycast);
+
+        /// <summary>
+        /// Attempts to add a new persistent raycast. The raycast will be updated automatically until
+        /// this subsystem is stopped or destroyed, or the the raycast is removed with
+        /// <see cref="RemoveRaycast(TrackableId)"/>.
+        /// </summary>
+        /// <param name="ray">A [ray](https://docs.unity3d.com/ScriptReference/Ray.html) relative to the session origin defining the raycast.</param>
+        /// <param name="estimatedDistance">The estimated distance to the raycast target. For example, an average
+        /// human height might be used to estimate the distance to the floor.</param>
+        /// <param name="raycast">The newly added raycast. All spatial data is relative to the session origin.</param>
+        /// <returns>`True` if the raycast was successfully added, or `false` otherwise.</returns>
+        public bool TryAddRaycast(Ray ray, float estimatedDistance, out XRRaycast raycast) => m_Provider.TryAddRaycast(ray, estimatedDistance, out raycast);
+
+        /// <summary>
+        /// Removes an existing raycast by its <see cref="TrackableId"/>.
+        /// </summary>
+        /// <param name="trackableId">The unique identifier for the raycast to remove.</param>
+        public void RemoveRaycast(TrackableId trackableId) => m_Provider.RemoveRaycast(trackableId);
 
         /// <summary>
         /// Casts <paramref name="ray"/> against trackables specified with <paramref name="trackableTypeMask"/>.
@@ -91,6 +138,57 @@ namespace UnityEngine.XR.ARSubsystems
             public virtual void Destroy() { }
 
             /// <summary>
+            /// Adds a new persistent raycast. Persistent raycasts should be updated automatically until this
+            /// provider is stopped or destroyed or the raycast is removed with
+            /// <see cref="RemoveRaycast(TrackableId)"/>.
+            /// </summary>
+            /// <param name="screenPoint">A position on the screen in normalized screen coordinates (0, 0)..(1, 1)</param>
+            /// <param name="estimatedDistance">The estimated distance to the raycast target.</param>
+            /// <param name="raycast">The newly added raycast. All spatial data should be reported relative to the session origin.</param>
+            /// <returns>`True` if the raycast was added; otherwise `false`.</returns>
+            public virtual bool TryAddRaycast(Vector2 screenPoint, float estimatedDistance, out XRRaycast raycast)
+            {
+                raycast = XRRaycast.defaultValue;
+                return false;
+            }
+
+            /// <summary>
+            /// Adds a new persistent raycast. Persistent raycasts should be updated automatically until this
+            /// provider is stopped or destroyed or the raycast is removed with
+            /// <see cref="RemoveRaycast(TrackableId)"/>.
+            /// </summary>
+            /// <param name="ray">A ray in session-space defining the raycast</param>
+            /// <param name="estimatedDistance">The estimated distance to the raycast target.</param>
+            /// <param name="raycast">The newly added raycast. All spatial data should be reported relative to the session origin.</param>
+            /// <returns>`True` if the raycast was added; otherwise `false`.</returns>
+            public virtual bool TryAddRaycast(Ray ray, float estimatedDistance, out XRRaycast raycast)
+            {
+                raycast = XRRaycast.defaultValue;
+                return false;
+            }
+
+            /// <summary>
+            /// Removes a raycast previously added with
+            /// <see cref="TryAddRaycast(UnityEngine.Vector2,float,out UnityEngine.XR.ARSubsystems.XRRaycast)"/>
+            /// or
+            /// <see cref="TryAddRaycast(UnityEngine.Ray,float,out UnityEngine.XR.ARSubsystems.XRRaycast)"/>
+            /// </summary>
+            /// <param name="trackableId">The unique identifier associated with the raycast to remove.</param>
+            public virtual void RemoveRaycast(TrackableId trackableId) { }
+
+            /// <summary>
+            /// Get the changes (arrays of added, updated, and removed) raycasts since the last call to this method.
+            /// </summary>
+            /// <param name="defaultRaycast">A default value for <see cref="XRRaycast"/>s. For backwards compatibility,
+            /// this should be used to initialized the returned
+            /// [NativeArray](https://docs.unity3d.com/ScriptReference/Unity.Collections.NativeArray_1.html)s.</param>
+            /// <param name="allocator">The [Allocator](https://docs.unity3d.com/ScriptReference/Unity.Collections.Allocator.html)
+            /// to use when allocating the returned [NativeArray](https://docs.unity3d.com/ScriptReference/Unity.Collections.NativeArray_1.html)s.</param>
+            /// <returns>Arrays of added, updated, and removed raycasts since the last call to this method. The
+            /// changes should be allocated using <paramref name="allocator"/>. The caller owns the memory.</returns>
+            public virtual TrackableChanges<XRRaycast> GetChanges(XRRaycast defaultRaycast, Allocator allocator) => default;
+
+            /// <summary>
             /// Performs a raycast from an arbitrary ray against the types
             /// specified by <paramref name="trackableTypeMask"/>. Results
             /// should be sorted by distance from the ray origin.
@@ -130,5 +228,9 @@ namespace UnityEngine.XR.ARSubsystems
         }
 
         Provider m_Provider;
+
+#if DEVELOPMENT_BUILD || UNITY_EDITOR
+        ValidationUtility<XRRaycast> m_ValidationUtility = new ValidationUtility<XRRaycast>();
+#endif
     }
 }